Finishes library (CI, Doctrine) updates. Other minor fixes.
Some changes to the Al/Be tracking sheet.
Various updates related to changes in Doctrine.

Generating yaml from the database and then using that to generate the
models no longer works properly in the current Doctrine. It fails
to produce the desired tables relations, which was the whole point
of doing it this way in the first place. Any future changes to the
schema will need to specify any table relations in the yaml file
at app/schema/schema.yaml and then use the doctrine command line tool
to generate the models from that.

Some fixes to form validation.