Allow numeric beaker names in ICP data submission form
Perry Spector noticed that submitting ICP data for
numerical beaker names would fail with an error message
stating that beakers "0, 1, 2" etc. were not found.
It turns out this is a quirk of array_merge that
duplicate numerical entries (even if they are strings)
would be replaced with new numeric keys starting from zero.
The function array_push does not have this issue
and so is a suitable replacement.