Add isotope ratios and errors to carrier tables.
Fixes #1.

This commit includes the new yaml schema, models updated to include
the isotope ratio columns, and a migration to the new schema. Users
should be able to migrate to the new schema by updating to the
newest codebase and running "app/doctrine migrate" from the base of
the project directory.

We add a script "generate-migrations-diff" to generate a migration
from an old yaml schema at app/schema/schema_old.yml to the new one
at app/schema/schema.yml. This works around a bug in the Doctrine1
cli that says it can't find a php or yml file in the schema dir.
We should seriously consider switching to another DBAL. At any rate,
running this script creates a migration to the new database schema.