Add controller for adding chemistry containers
The Container controller allows the user to add, edit and view the
names of split beakers and dissolution bottles.

Added containers to the main navigation bar.

Added a container navbar that changes depending on which containers
are being viewed.

I made some adjustments to the _is_unique callback in the base
controller so that it can cope with IDs placed in arbitrary
URI segments. The uri segment with the ID (if known) is specified
in the form validation rule as an integer.

Also fixed some minor CSS issues that were also affected the new
controller.

Deleting containers is currently not possible in the interface.
We may want to add this, but we will need to make sure that no
analyses use the container before allowing a delete.