Add form validation for carriers
BeCarrier and AlCarrier share validation rules from BaseCarrier that are
merged into their form validation arrays.

Currently in_use is only treated as a string and not used for anything
else except by external clients. Eventually we will want to make this a
checkbox and ensure that it is only 'y' or 'n'. This should rightly be a
boolean but I will need to check with legacy users who have written
scripts against this database schema.