Avoid printing trailing zeros by casting to float
Cast to (float) everywhere that values are computed or drawn from the
database. This eliminates the trailing zeros that we now see since
changing the MySQL data type of our measured values from float to
double. In PHP, float and double are synonymous, so we don't lose any
precision.

In a few cases numbers are rounded and/or run through sprintf prior to
being passed to the view, so I did not include casts for these echo
statements. Since that is display logic, we should consider moving those
commands from the controllers into the views in the future.