Remove schema declarations that are defaults
Many of the declarations in the yaml schema are the default and make the
schema unnecessarily long. Here we remove these default declarations and
make the corresponding changes in the generated model classes using
`doctrine generate-models-yaml`. Subsequently running `doctrine
generate-sql` results in no changes to the generated SQL.

This creates migration version 30.