Use default size for integer columns (migr. 31)
We have a variety of integer sizes in the schema, but some may not be
large enough in the future. This changes the schema to use the default
integer size rather than specifying any smaller sizes so that size
should not be an issue.

While we're at it, we remove empty parens after string fields in the yaml.

Adds migration 31.

Because this changes fields involved in foreign key constraints, in
order to allow the migration to proceed, the generated code has been
edited to first remove the foreign key constraints and afterwards
recreate them.

Two foreign key constraints, `analysis_diss_bottle_id_diss_bottle_id`
and `batch_be_carrier_id_be_carrier_id`, had not previously been
included in the migrations (due to issue #3), but may have been created
in recent new installations. After running this migration they should be
created properly on all installations.