add_split_weights: Fix setting beaker sequence
Currently setting the split beaker sequence is broken unless the split
beakers are named like as they are at UW (prefixed by AB). To fix it,
the select options are now sorted by the beaker label and setting the
beaker sequence is done by walking through the option elements following
the first one selected until the next one in the sequence is found. It
is then set to selected. It no longer reads labels of any of the beakers.

Additionally, we move the beaker sequence javascript into its own file
and remove an ancient onclick call that did not do anything.