Add setup script that sets migration_version
It creates the database, creates the tables, and sets the migration
version to the current value. Previously an unset migration_version
would cause problems when running a migration for the first time and
the user would need to first set the migration version manually.

The migration version in this script will need to be bumped every time
we add a new migration.